Implemented multi-select, i.e. clicking on an object to select it will not clear the current selection first -- merely toggle it.
This required some big changes and a hack or two, for one thing
when clicking on an object, it's selection status is changed when
the the mouse button is _released_ (or user tries to drag it).

Also to allow dragging individual objects, after any objects are
moved we will clear the selection when an unselected object is
clicked on (to select it).

Removed the dummy 'CANVAS' object kind, we now properly test
whether we are drawing a selbox by the isSelboxActive() call.