LideDef panel: changed layout of textures in each SideDef section. The default order is now LOWER / UPPER / RAIL.
Also, single-sided linedefs show the texture in the "LOWER" position,
since there is no longer a spot for a "middle" texture.

The main benefit of this is when multiple linedefs are selected which
is a mix of one-sided and two-sided lines, the user can safely change
the "LOWER" (or LOWER + UPPER) on all the linedefs without ending up
with unwanted railing textures.