handle a very rare assertion failure in m_strings.cc more gracefully.
Note: I would normally look for the root cause, but internalised
strings are used a lot throughout the code, so it is impossible to
know where the real problem is unless you can trigger the issue
while using a debugger.  Hence added a workaround here for the
sake of robustness.