EditLump : treat "SCRIPTS" as a invalid lump (prevent editing).
Normally "SCRIPTS" is a level lump, appearing directly after
the "BEHAVIOR" lump, and is completely optional.  There are some
wad files which contain a global "SCRIPTS" lump (not attached to
any particular level).

Trying to support such a duality leads to problems, especially
when the user creates a global "SCRIPTS" lump, but it ends up
directly after a level inside the wad, and hence suddenly looks
like a level lump (not global anymore).  So the user may find that
their edited lump has "disappeared" when next trying to edit it,
which is totally unacceptable.

Investigations show that source ports (esp. ZDoom) never read a
global "SCRIPTS" lump, it is just a way for modders to bundle
the source of their ACS scripts (as a courtesy).  Hence I think
it is better to err on the side of safety, and only allow the
creation/editing of level-local "SCRIPTS" lumps.