3D View : worked on rewriting the wall sorting code.
The previous code had some problems, especially when the wall
list contained many DrawWalls at the same depth, i.e. sprites.

This commit is a preliminary fix to the QuickSort algorithm,
in particular the case when everything was >= pivot, the old
code did not ensure the pivot was moved to the beginning.