Game def parser : support new "clear" command.
This command takes one or more keywords, such as "lines" or
"sectors", and clears the corresponding set of types.

Use this keyword in the ZDoom port definitions to clear
out the DOOM line and sector specials when the map format
is HEXEN, to better support "Doom in Hexen Format" maps.