wad code : in ReadDirectory, detect invalid lumps.
In other words, detect lumps with where the start or length
are negative or beyond the end of the wad file.

When found, we log a warning message and clear the lump
(make it empty), since an invalid position would lead to an
error or assertion failure later on (e.g. when saving).