canvas : speed up drawing selection when it has many sectors.
This affected all OSes (not just X windows), especially in
maps with a huge number of linedefs.  The slowdown was caused
by iterating over the selected sectors one-by-one, and for
each sector we iterated over ALL the linedefs.

This commit uses a better method when the selection mode is
OBJ_SECTORS and the number of sectors is large.  It performs
just a single iteration over the linedefs.