Makefile : fixed not to use "fltk-config --libs".
That's because "--ldflags" provides the libraries for *dynamic*
linking, but "--libs" provides libs for *static* linking, and we
should not have both.

Also separated important flags to CXX_DIALECT variable, and made
the base CXXFLAGS easier to override (though the CXX_DIALECT flags
are always added, since they define important properties and the
code may not work properly without them).