gl-canvas : added DeleteContext() to delete the current OpenGL context.
This will be called whenever all the GL textures need to be dumped,
such as when Eureka reloads all textures, or when changing the gamma.

This commit also implements invalidating the 'gl_tex' field of all
currently loaded images (Img_c class), which will force those images
to be uploaded to the new context on first use.