bsp : simplified cut-list and wall-tips to not care about sectors.
There really isn't the nodes builder can do when sector mismatches
occur, and a few hundred minor warning messages is not particularly
helpful for a user.

Hence updated the code to only worry about whether the position
in question is OPEN (a sector) or CLOSED (void or along a linedef).