in main() function, call Main_LoadResources() *after* LoadLevel().
This is for the initial level loaded.  It matches the behavior
of the other places where LoadLevel() is called, especially the
CMD_OpenMap() function.  Thankfully nothing in the LoadLevel()
code depends on any game or port definitions -- it is fairly raw.

The new order is needed for UDMF so that we can nicely handle
cases where the namespace says "ZDoom" or "Eternity" but our
current port is not UDMF capable (like "vanilla").