udmf : removed UDMF_NAMESPACE checks from ZDoom and Eternity port defs.
That's because we're only going to support the engine-specific
namespace for UDMF maps, and not the namespaces reserved in the
UDMF spec ("Doom", "Heretic", "Hexen" and "Strife").

Also changed "udmf_namespace" to require a single name, which
will be the engine-specific namespace.  Just using the raw port
name probably would have worked, but this provides the nicely
capitalized name.