bindings.cfg : unbound plain L/S/T/V keys in 3D mode.
Since the 3D mode highlight and selection now depends on the
current editing mode, it is important to be able to switch
modes using the existing "EditMode" key bindings.

Unfortunately that means the WASD keys can no longer be used
for navigation, hence unbound those keys from the 3D mode.

The plain L and T keys were bound in 3D mode to toggle the
lighting and texturing.  These now have the META modifier,
which means they can be accessed by pressing ';' first.

The plain V key was bound to "drop to floor" command, so that
has also added the META modifier.  The existing G key which
toggles gravity on/off serves a similar purpose anyway.