added safety checks to prevent infinite FatalError() loops.
I just discovered this is possible by a failing SYS_ASSERT in the
UI_Canvas drawing code.  Turns out that the dialog we use to show
the error message can re-enter the main window's draw() method!

Two safety checks are now in place.  Firstly, mark when a fatal
error has occurred and if FatalError() gets called again then
abort the program *without* showing a dialog.  Secondly, check
the fatal error variable in UI_MainWindow's draw() and handle()
methods and only continue when it is clear.