some dead-end work to support texture copy/paste in Find/Replace panel.
The reason it is a dead end is that the UI_Pic showing the flat
or texture does not get the keyboard focus when highlighted or
selected, and hence CTRL-C and CTRL-V keypresses are usually
sent to the FL_Input (which do some copy/paste for text).

Actually the Defaults panel and Sidedef sections of the LineDef
panel are not immune to this problem, but it is less noticeable
since they don't focus the FL_Input widgets by default.

I think the only way this can be completely fixed is to create a
custom widget containing both an input box and a UI_Pic, and to
assign focus to the proper one when needed.  But that is a very
heavy solution for a small bit of functionality....