fixed cases where a texture/flat/thing UI_Pic could be stuck as highlighted.
The bug occurred because the UI_Pic widget relied on an FL_LEAVE event
to unhighlight itself when the mouse moves outside, but other code
could make the UI_Pic (or the panel it's in) hidden and hence it
never received that FL_LEAVE event.

A consequence of this bug was that Copy'n'Paste in the 3D view could
strangely stop working, even though the mouse is nowhere near the panel,
since a highlighted or selected pic in the panel gets priority.