canvas : renamed "highlight" field to "last_highlight", but don't access it.
This field is ONLY there to detect when a change has occurred and
prevent unnecessary redraws.  Previous code was also accessing this
value instead of "edit.highlight", which is likely the cause of the
crashes people were getting when the code used an invalid obj num.

Made a similar change for "split_ld" field, renamed it "last_splitter"
and access the "edit.split_xxx" values instead of this field.