3D view : removed "current_hl" field from Render_View_t.
This field had the potential to get out-of-date, similar to what
just happened with "highlight" field in UI_Canvas class.

Instead of that, we remember the mouse coordinate after a motion
event, and Render3D_UpdateHighlight() computes the new highlight
object and map coordinate.  This is consistent with what the 2D
UpdateHighlight() function does.