Honor negative patch offsets in textures when using ZDoom
This is an odd one. Yes, I'm really removing a "fix" commented with a
timestamp from 1998! Almost 22 years ago to the day.

The comment said that the original DOOM weirdly did not honor negative
y-offsets set on texture patches. Perhaps this was a bug but we'll
probably never know. TEKWALL1 is one of the textures mentioned.

I was trying to work out why my own textures didn't look right so the
first thing I did was remove this fix. It did not fix the issue but
strangely, it did not affect TEKWALL1 either.

I then found a similar line in w_loadpic's DrawColumn. Adjusting that
did the trick for my own patches but had I broken TEKWALL1?

It did indeed look different but when I compared it against GZDoom, I
found that it matched the new rendering instead. Had I actually
*fixed* TEKWALL1? I checked SLADE and Ultimate Doom Builder and those
matched too. Perhaps I had.

That fix had been there for so long though and that was bugging
me. Surely it had been added for a reason. I then checked Chocolate
Doom. To my surprise, this matched the previous rendering. Finally, I
checked PrBoom-Plus. That was the same.

So on the one hand, we have GZDoom, SLADE, and UDB showing the logical
rendering and on the other hand, we have Chocolate Doom and
PrBoom-Plus presumably matching the original game. Which is correct?

I asked this question on the GZDoom Discord server but I didn't get a
reply. Perhaps they didn't care. I did some archaeological digging and
found an ancient discussion about this in the ZDoom forums and then
found code relating to this still in the engine today, in d_main.cpp
no less.

Their fix is only applied to SKY1 and BIGDOOR7 specifically
though. With SKY1, you'd be forgiven for not noticing the
difference. BIGDOOR7 is confusing because it says it has a negative
y-offset but it only has a negative x-offset from what I can see. I've
compared it across GZDoom, Chocolate Doom, and Eureka, and they all
look the same to me.

Since GZDoom seemingly isn't going to change anything here and since
custom patches should probably be rendered correctly anyway, I have
added a port feature to honor negative patch offsets. If you want to
see the difference for yourself, change the port in the settings while
viewing TEKWALL1. I have not made SKY1 a special case as it doesn't
seem worth it.